Allow for ESM webpack modules
