--- env.d.ts
@@ -1,1 +1,0 @@
-/// <reference types="./packages/types/src/index.d.ts" />
