+++ packages/core-extensions/src/common/flux.ts
@@ -6,7 +6,7 @@ const findFlux = ["useStateFromStores:function"];
 export const flux: ExtensionWebpackModule = {
   dependencies: [{ ext: "spacepack", id: "spacepack" }, ...findFlux],
   run: (module, exports, require) => {
-    const spacepack = require("spacepack_spacepack");
+    const spacepack = require("spacepack_spacepack").spacepack;
     const Flux = spacepack.findByCode(...findFlux)[0].exports;
     module.exports = Flux as CommonFlux;
   }
