+++ packages/core-extensions/src/common/http.ts
@@ -5,7 +5,7 @@ const findHTTP = ["get", "put", "V8APIError"];
 export const http: ExtensionWebpackModule = {
   dependencies: [{ ext: "spacepack", id: "spacepack" }],
   run: (module, exports, require) => {
-    const spacepack = require("spacepack_spacepack");
+    const spacepack = require("spacepack_spacepack").spacepack;
     const HTTP = spacepack.findByExports(...findHTTP)[0].exports;
     module.exports = HTTP.ZP ?? HTTP.Z;
   }
